rename TokenAuth -> StaticTokenAuthHandler, move to node.rs
